requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
certifi
plaid-python==5.0.0